+++ coreax/solvers/__init__.py
@@ -27,6 +27,7 @@ from coreax.solvers.coresubset import (
     ConditionalKernelHerdingState,
     HerdingState,
     JointKernelHerding,
+    JointRPCholesky,
     KernelHerding,
     RandomSample,
     RPCholesky,
@@ -50,5 +51,6 @@ __all__ = [
     "PaddingInvariantSolver",
     "ConditionalKernelHerding",
     "ConditionalKernelHerdingState",
+    "JointRPCholesky",
     "JointKernelHerding",
 ]