+++ coreax/solvers/coresubset.py
@@ -478,37 +478,49 @@ class RandomSample(CoresubsetSolver[_Data, None], ExplicitSizeSolver):
             raise
 
 
-class RPCholesky(CoresubsetSolver[_Data, RPCholeskyState], ExplicitSizeSolver):
+class _GenericDataRPCholesky(
+    CoresubsetSolver[Union[_Data, _SupervisedData], RPCholeskyState], ExplicitSizeSolver
+):
     r"""
-    Randomly Pivoted Cholesky - an explicitly sized coresubset refinement solver.
+    An implementation of Randomly Pivoted Cholesky handling (un)supervised data types.
 
     Solves the coresubset problem by taking a stochastic, iterative, and greedy approach
     to approximating the Gramian of a given kernel, evaluated on the original dataset.
 
+    .. note::
+        :class:`_GenericDataRPCholesky` should not be used directly, if compressing
+        unsupervised :class:`~coreax.data.Data`, use :class:`RPCholesky`, if compressing
+        :class:`~coreax.data.SupervisedData`, use :class:`JointRPCholesky`.
+
     :param coreset_size: The desired size of the solved coreset
     :param random_key: Key for random number generation
     :param kernel: :class:`~coreax.kernel.Kernel` instance implementing a kernel
-        function :math:`k: \mathbb{R}^d \times \mathbb{R}^d \rightarrow \mathbb{R}`
+        function :math:`k: \mathbb{R}^d \times \mathbb{R}^d \rightarrow \mathbb{R}` if
+        compressing unsupervised :class:`~coreax.data.Data`, else
+        :class:`~coreax.kernel.TensorProduct` instance if compressing
+        :class:`~coreax.data.SupervisedData`
     :param unique: If each index in the resulting coresubset should be unique
     """
 
     random_key: KeyArrayLike
-    kernel: Kernel
+    kernel: Union[Kernel, TensorProductKernel]
     unique: bool = True
 
     def reduce(
-        self, dataset: _Data, solver_state: Optional[RPCholeskyState] = None
-    ) -> tuple[Coresubset[_Data], RPCholeskyState]:
+        self,
+        dataset: Union[Data, SupervisedData],
+        solver_state: Optional[RPCholeskyState] = None,
+    ) -> tuple[Coresubset[Union[Data, SupervisedData]], RPCholeskyState]:
         """
-        Reduce 'dataset' to a :class:`~coreax.coreset.Coresubset` with 'RPCholesky'.
+        Reduce 'dataset' to a :class:`~coreax.coreset.Coresubset`.
 
-        This is done by first computing the kernel Gram matrix of the original data, and
-        isolating the diagonal of this. A 'pivot point' is then sampled, where sampling
-        probabilities correspond to the size of the elements on this diagonal. The
-        data-point corresponding to this pivot point is added to the coreset, and the
-        diagonal of the Gram matrix is updated to add a repulsion term of sorts -
-        encouraging the coreset to select a range of distinct points in the original
-        data. The pivot sampling and diagonal updating steps are repeated until
+        This is done by first computing the (tensor-product) kernel Gram matrix of the
+        original data, and isolating the diagonal of this. A 'pivot point' is then
+        sampled, where sampling probabilities correspond to the size of the elements on
+        this diagonal. The data-point corresponding to this pivot point is added to the
+        coreset, and the diagonal of the Gram matrix is updated to add a repulsion term
+        of sorts - encouraging the coreset to select a range of distinct points in the
+        original data. The pivot sampling and diagonal updating steps are repeated until
         :math:`M` points have been selected.
 
         :param dataset: The dataset to reduce to a coresubset
@@ -516,11 +528,28 @@ class RPCholesky(CoresubsetSolver[_Data, RPCholeskyState], ExplicitSizeSolver):
             expensive intermediate solution step values.
         :return: a refined coresubset and relevant intermediate solver state information
         """
-        x = dataset.data
-        if solver_state is None:
-            gramian_diagonal = jax.vmap(self.kernel.compute_elementwise)(x, x)
-        else:
+        # Setup the class to deal with supervised or unsupervised data
+        if solver_state is not None:
             gramian_diagonal = solver_state.gramian_diagonal
+        if isinstance(self.kernel, TensorProductKernel) and isinstance(
+            dataset, SupervisedData
+        ):
+            x, y = dataset.data, dataset.supervision
+            if solver_state is None:
+                gramian_diagonal = jax.vmap(self.kernel.compute_elementwise)(
+                    (x, y), (x, y)
+                )
+        elif isinstance(self.kernel, Kernel):
+            x = dataset.data
+            if solver_state is None:
+                gramian_diagonal = jax.vmap(self.kernel.compute_elementwise)(x, x)
+        else:
+            raise ValueError(
+                "Invalid combination of 'kernel' and 'dataset'; if compressing"
+                + " 'SupervisedData', one must pass a 'TensorProductKernel', if"
+                + " compressing 'Data', one must pass a child of 'Kernel'."
+            )
+
         initial_coresubset = _initial_coresubset(0, self.coreset_size, dataset)
         coreset_indices = initial_coresubset.unweighted_indices
         num_data_points = len(x)
@@ -528,7 +557,7 @@ class RPCholesky(CoresubsetSolver[_Data, RPCholeskyState], ExplicitSizeSolver):
         def _greedy_body(
             i: int, val: tuple[Array, Array, Array]
         ) -> tuple[Array, Array, Array]:
-            """RPCholesky iteration - Algorithm 1 of :cite:`chen2023randomly`."""
+            """RPCholesky Iteration - Algorithm 1 of :cite:`chen2023randomly`."""
             residual_diagonal, approximation_matrix, coreset_indices = val
             key = jr.fold_in(self.random_key, i)
             pivot_point = jr.choice(
@@ -536,10 +565,20 @@ class RPCholesky(CoresubsetSolver[_Data, RPCholeskyState], ExplicitSizeSolver):
             )
             updated_coreset_indices = coreset_indices.at[i].set(pivot_point)
             # Remove overlap with previously chosen columns
-            g = (
-                self.kernel.compute(x, x[pivot_point])
-                - (approximation_matrix @ approximation_matrix[pivot_point])[:, None]
-            )
+            if isinstance(self.kernel, TensorProductKernel):
+                g = (
+                    self.kernel.compute((x, y), (x[pivot_point], y[pivot_point]))
+                    - (approximation_matrix @ approximation_matrix[pivot_point])[
+                        :, None
+                    ]
+                )
+            else:
+                g = (
+                    self.kernel.compute(x, x[pivot_point])
+                    - (approximation_matrix @ approximation_matrix[pivot_point])[
+                        :, None
+                    ]
+                )
             updated_approximation_matrix = approximation_matrix.at[:, i].set(
                 jnp.ravel(g / jnp.sqrt(g[pivot_point]))
             )
@@ -564,6 +603,134 @@ class RPCholesky(CoresubsetSolver[_Data, RPCholeskyState], ExplicitSizeSolver):
         _, _, updated_coreset_indices = output_state
         updated_coreset = Coresubset(updated_coreset_indices, dataset)
         return updated_coreset, RPCholeskyState(gramian_diagonal)
+
+
+class RPCholesky(CoresubsetSolver[_Data, RPCholeskyState], ExplicitSizeSolver):
+    r"""
+    Randomly Pivoted Cholesky - an explicitly sized coresubset refinement solver.
+
+    Solves the coresubset problem by taking a stochastic, iterative, and greedy approach
+    to approximating the Gramian of a given kernel, evaluated on the original dataset.
+
+    .. note::
+        :class:`RPCholesky` is suitable for compressing unsupervised
+        :class:`~coreax.data.Data`, use :class:`JointRPCholesky`, if compressing
+        :class:`~coreax.data.SupervisedData`.
+
+    :param coreset_size: The desired size of the solved coreset
+    :param random_key: Key for random number generation
+    :param kernel: :class:`~coreax.kernel.Kernel` instance implementing a kernel
+        function :math:`k: \mathbb{R}^d \times \mathbb{R}^d \rightarrow \mathbb{R}`
+    :param unique: If each index in the resulting coresubset should be unique
+    """
+
+    random_key: KeyArrayLike
+    kernel: Kernel
+    unique: bool = True
+
+    def reduce(
+        self, dataset: Data, solver_state: Optional[RPCholeskyState] = None
+    ) -> tuple[Coresubset[Data], RPCholeskyState]:
+        """
+        Reduce `dataset` to a :class:`~coreax.coreset.Coresubset` with `RPCholesky`.
+
+        This is done by first computing the kernel Gram matrix of the original data, and
+        isolating the diagonal of this. A 'pivot point' is then sampled, where sampling
+        probabilities correspond to the size of the elements on this diagonal. The
+        data-point corresponding to this pivot point is added to the coreset, and the
+        diagonal of the Gram matrix is updated to add a repulsion term of sorts -
+        encouraging the coreset to select a range of distinct points in the original
+        data. The pivot sampling and diagonal updating steps are repeated until
+        :math:`M` points have been selected.
+
+        :param dataset: The dataset to reduce to a coresubset
+        :param solver_state: Solution state information, primarily used to cache
+            expensive intermediate solution step values.
+        :return: a refined coresubset and relevant intermediate solver state information
+        """
+        unsupervised_solver = _GenericDataRPCholesky(
+            coreset_size=self.coreset_size,
+            random_key=self.random_key,
+            kernel=self.kernel,
+            unique=self.unique,
+        )
+        return unsupervised_solver.reduce(dataset, solver_state)
+
+
+class JointRPCholesky(
+    CoresubsetSolver[_SupervisedData, RPCholeskyState], ExplicitSizeSolver
+):
+    r"""
+    Joint Randomly Pivoted Cholesky - an explicitly sized coresubset refinement solver.
+
+    Solves the coresubset problem by taking a stochastic, iterative, and greedy approach
+    to approximating the Gramian of a tensor-product kernel, evaluated on the original
+    supervised dataset.
+
+    .. note::
+        :class:`JointRPCholesky` is suitable for compressing
+        :class:`~coreax.data.SupervisedData`, use :class:`RPCholesky`, if compressing
+        unsupervised :class:`~coreax.data.Data`.
+
+    :param coreset_size: The desired size of the solved coreset
+    :param random_key: Key for random number generation
+    :param feature_kernel: :class:`~coreax.kernel.Kernel` instance implementing a kernel
+        function :math:`k: \mathbb{R}^d \times \mathbb{R}^d \rightarrow \mathbb{R}` on
+        the feature space
+    :param response_kernel: :class:`~coreax.kernel.Kernel` instance implementing a
+        kernel function :math:`k: \mathbb{R}^p \times \mathbb{R}^p \rightarrow
+        \mathbb{R}` on the response space
+    :param unique: If each index in the resulting coresubset should be unique
+    """
+
+    random_key: KeyArrayLike
+    kernel: TensorProductKernel
+    unique: bool = True
+
+    def __init__(
+        self,
+        coreset_size: int,
+        random_key: KeyArrayLike,
+        feature_kernel: Kernel,
+        response_kernel: Kernel,
+        unique: bool = True,
+    ):
+        """Initialise JointRPCholesky class and build tensor-product kernel."""
+        ExplicitSizeSolver.__init__(self, coreset_size)
+        self.random_key = random_key
+        self.kernel = TensorProductKernel(
+            feature_kernel=feature_kernel,
+            response_kernel=response_kernel,
+        )
+        self.unique = unique
+
+    def reduce(
+        self, dataset: SupervisedData, solver_state: Optional[RPCholeskyState] = None
+    ) -> tuple[Coresubset[SupervisedData], RPCholeskyState]:
+        """
+        Reduce `dataset` to :class:`~coreax.coreset.Coresubset` with `JointRPCholesky`.
+
+        This is done by first computing the tensor-product kernel Gram matrix of the
+        original supervised data pairs, and isolating the diagonal of this. A
+        'pivot point' is then sampled, where sampling probabilities correspond to the
+        size of the elements on this diagonal. The data-point corresponding to this
+        pivot point is added to the coreset, and the diagonal of the Gram matrix is
+        updated to add a repulsion term of sorts - encouraging the coreset to select a
+        range of distinct pairs in the original data. The pivot sampling and diagonal
+        updating steps are repeated until :math:`M` pairs have been selected.
+
+        :param dataset: The dataset to reduce to a coresubset
+        :param solver_state: Solution state information, primarily used to cache
+            expensive intermediate solution step values.
+        :return: a refined coresubset and relevant intermediate solver state information
+        """
+        supervised_solver = _GenericDataRPCholesky(
+            coreset_size=self.coreset_size,
+            random_key=self.random_key,
+            kernel=self.kernel,
+            unique=self.unique,
+        )
+        return supervised_solver.reduce(dataset, solver_state)
 
 
 class SteinThinning(