+++ tests/unit/test_solvers.py
@@ -45,6 +45,7 @@ from coreax.kernel import (
 from coreax.solvers import (
     ConditionalKernelHerding,
     JointKernelHerding,
+    JointRPCholesky,
     KernelHerding,
     MapReduce,
     RandomSample,
@@ -102,7 +103,7 @@ class SolverTest:
         self,
         request: pytest.FixtureRequest,
         solver_factory: Union[type[Solver], jtu.Partial],
-    ) -> _ReduceProblem:
+    ) -> Union[_ReduceProblem, _SupervisedReduceProblem]:
         """
         Pytest fixture that returns a problem dataset and the expected coreset.
 
@@ -116,7 +117,9 @@ class SolverTest:
         return _ReduceProblem(Data(dataset), solver, expected_coreset)
 
     def check_solution_invariants(
-        self, coreset: Coreset, problem: Union[_RefineProblem, _ReduceProblem]
+        self,
+        coreset: Coreset,
+        problem: Union[_RefineProblem, _ReduceProblem, _SupervisedReduceProblem],
     ) -> None:
         """
         Check that a coreset obeys certain expected invariant properties.
@@ -157,7 +160,7 @@ class SolverTest:
     def test_reduce(
         self,
         jit_variant: Callable[[Callable], Callable],
-        reduce_problem: _ReduceProblem,
+        reduce_problem: Union[_ReduceProblem, _SupervisedReduceProblem],
         use_cached_state: bool,
     ) -> None:
         """
@@ -270,7 +273,9 @@ class ExplicitSizeSolverTest(SolverTest):
 
     @override
     def check_solution_invariants(
-        self, coreset: Coreset, problem: Union[_RefineProblem, _ReduceProblem]
+        self,
+        coreset: Coreset,
+        problem: Union[_RefineProblem, _ReduceProblem, _SupervisedReduceProblem],
     ) -> None:
         super().check_solution_invariants(coreset, problem)
         solver = problem.solver
@@ -304,7 +309,7 @@ class ExplicitSizeSolverTest(SolverTest):
             assert solver.coreset_size == int(coreset_size)
 
     def test_reduce_oversized_coreset_size(
-        self, reduce_problem: _ReduceProblem
+        self, reduce_problem: Union[_ReduceProblem, _SupervisedReduceProblem]
     ) -> None:
         """
         Check error is raised if 'coreset_size' is too large.
@@ -318,6 +323,29 @@ class ExplicitSizeSolverTest(SolverTest):
         )
         with pytest.raises(ValueError, match=re.escape(self.OVERSIZE_MSG)):
             modified_solver.reduce(dataset)
+
+
+class TestRandomSample(ExplicitSizeSolverTest):
+    """Test cases for :class:`coreax.solvers.coresubset.RandomSample`."""
+
+    @override
+    def check_solution_invariants(
+        self,
+        coreset: Coreset,
+        problem: Union[_RefineProblem, _ReduceProblem, _SupervisedReduceProblem],
+    ) -> None:
+        super().check_solution_invariants(coreset, problem)
+        solver = cast(RandomSample, problem.solver)
+        if solver.unique:
+            _, counts = jnp.unique(coreset.nodes.data, return_counts=True)
+            assert max(counts) <= 1
+
+    @override
+    @pytest.fixture(scope="class")
+    def solver_factory(self) -> jtu.Partial:
+        coreset_size = self.shape[0] // 10
+        key = jr.fold_in(self.random_key, self.shape[0])
+        return jtu.Partial(RandomSample, coreset_size=coreset_size, random_key=key)
 
 
 class TestKernelHerding(RefinementSolverTest, ExplicitSizeSolverTest):
@@ -382,18 +410,18 @@ class TestKernelHerding(RefinementSolverTest, ExplicitSizeSolverTest):
 class TestJointKernelHerding(RefinementSolverTest, ExplicitSizeSolverTest):
     """Test cases for :class:`coreax.solvers.coresubset.JointKernelHerding`."""
 
-    additional_key = jr.key(0)
+    feature_kernel = SquaredExponentialKernel()
+    response_kernel = PCIMQKernel()
 
     @override
     @pytest.fixture(scope="class")
     def solver_factory(self) -> jtu.Partial:
-        kernel = SquaredExponentialKernel()
         coreset_size = self.shape[0] // 10
         return jtu.Partial(
             JointKernelHerding,
             coreset_size=coreset_size,
-            feature_kernel=kernel,
-            response_kernel=kernel,
+            feature_kernel=self.feature_kernel,
+            response_kernel=self.response_kernel,
         )
 
     @override
@@ -402,16 +430,17 @@ class TestJointKernelHerding(RefinementSolverTest, ExplicitSizeSolverTest):
         self,
         request: pytest.FixtureRequest,
         solver_factory: Union[type[Solver], jtu.Partial],
-    ) -> _ReduceProblem:
+    ) -> _SupervisedReduceProblem:
         if request.param == "random":
-            features = jr.uniform(self.random_key, self.shape)
-            responses = jr.uniform(self.additional_key, self.shape)
+            data_key, supervision_key = jr.split(self.random_key)
+            data = jr.uniform(data_key, self.shape)
+            supervision = jr.uniform(supervision_key, self.shape)
             solver = solver_factory()
             expected_coreset = None
         else:
             raise ValueError("Invalid fixture parametrization")
-        return _ReduceProblem(
-            SupervisedData(features, responses), solver, expected_coreset
+        return _SupervisedReduceProblem(
+            SupervisedData(data, supervision), solver, expected_coreset
         )
 
     def test_herding_state(self, reduce_problem: _SupervisedReduceProblem) -> None:
@@ -422,26 +451,16 @@ class TestJointKernelHerding(RefinementSolverTest, ExplicitSizeSolverTest):
         expected_state = HerdingState(solver.kernel.gramian_row_mean(dataset))
         assert eqx.tree_equal(state, expected_state)
 
-
-class TestRandomSample(ExplicitSizeSolverTest):
-    """Test cases for :class:`coreax.solvers.coresubset.RandomSample`."""
-
-    @override
-    def check_solution_invariants(
-        self, coreset: Coreset, problem: Union[_RefineProblem, _ReduceProblem]
-    ) -> None:
-        super().check_solution_invariants(coreset, problem)
-        solver = cast(RandomSample, problem.solver)
-        if solver.unique:
-            _, counts = jnp.unique(coreset.nodes.data, return_counts=True)
-            assert max(counts) <= 1
-
-    @override
-    @pytest.fixture(scope="class")
-    def solver_factory(self) -> jtu.Partial:
-        coreset_size = self.shape[0] // 10
-        key = jr.fold_in(self.random_key, self.shape[0])
-        return jtu.Partial(RandomSample, coreset_size=coreset_size, random_key=key)
+    def test_non_uniqueness(self, reduce_problem: _SupervisedReduceProblem) -> None:
+        """Test that setting `unique` to be false produces no errors."""
+        dataset, _, _ = reduce_problem
+        solver = JointKernelHerding(
+            coreset_size=10,
+            feature_kernel=self.feature_kernel,
+            response_kernel=self.response_kernel,
+            unique=False,
+        )
+        solver.reduce(dataset)
 
 
 class TestRPCholesky(ExplicitSizeSolverTest):
@@ -449,7 +468,9 @@ class TestRPCholesky(ExplicitSizeSolverTest):
 
     @override
     def check_solution_invariants(
-        self, coreset: Coreset, problem: Union[_RefineProblem, _ReduceProblem]
+        self,
+        coreset: Coreset,
+        problem: Union[_RefineProblem, _ReduceProblem, _SupervisedReduceProblem],
     ) -> None:
         """Check functionality of 'unique' in addition to the default checks."""
         super().check_solution_invariants(coreset, problem)
@@ -480,6 +501,91 @@ class TestRPCholesky(ExplicitSizeSolverTest):
         expected_state = RPCholeskyState(gramian_diagonal)
         assert eqx.tree_equal(state, expected_state)
 
+    def test_non_uniqueness(self, reduce_problem: _ReduceProblem) -> None:
+        """Test that setting `unique` to be false produces no errors."""
+        dataset, _, _ = reduce_problem
+        solver = RPCholesky(
+            random_key=self.random_key,
+            coreset_size=10,
+            kernel=PCIMQKernel(),
+            unique=False,
+        )
+        solver.reduce(dataset)
+
+
+class TestJointRPCholesky(ExplicitSizeSolverTest):
+    """Test cases for :class:`coreax.solvers.coresubset.JointRPCholesky`."""
+
+    feature_kernel = SquaredExponentialKernel()
+    response_kernel = PCIMQKernel()
+
+    @override
+    @pytest.fixture(scope="class")
+    def solver_factory(self) -> jtu.Partial:
+        coreset_size = self.shape[0] // 10
+        return jtu.Partial(
+            JointRPCholesky,
+            coreset_size=coreset_size,
+            random_key=self.random_key,
+            feature_kernel=self.feature_kernel,
+            response_kernel=self.response_kernel,
+        )
+
+    @override
+    @pytest.fixture(params=["random"], scope="class")
+    def reduce_problem(
+        self,
+        request: pytest.FixtureRequest,
+        solver_factory: Union[type[Solver], jtu.Partial],
+    ) -> _SupervisedReduceProblem:
+        if request.param == "random":
+            data_key, supervision_key = jr.split(self.random_key)
+            data = jr.uniform(data_key, self.shape)
+            supervision = jr.uniform(supervision_key, self.shape)
+            solver = solver_factory()
+            expected_coreset = None
+        else:
+            raise ValueError("Invalid fixture parametrization")
+        return _SupervisedReduceProblem(
+            SupervisedData(data, supervision), solver, expected_coreset
+        )
+
+    @override
+    def check_solution_invariants(
+        self,
+        coreset: Coreset,
+        problem: Union[_RefineProblem, _ReduceProblem, _SupervisedReduceProblem],
+    ) -> None:
+        """Check functionality of 'unique' in addition to the default checks."""
+        super().check_solution_invariants(coreset, problem)
+        solver = cast(JointRPCholesky, problem.solver)
+        if solver.unique:
+            _, counts = jnp.unique(coreset.nodes.data, return_counts=True)
+            assert max(counts) <= 1
+
+    def test_rpcholesky_state(self, reduce_problem: _SupervisedReduceProblem) -> None:
+        """Check that the cached JointRPCholesky state is as expected."""
+        dataset, solver, _ = reduce_problem
+        solver = cast(JointRPCholesky, solver)
+        _, state = solver.reduce(dataset)
+
+        x, y = dataset.data, dataset.supervision
+        gramian_diagonal = jax.vmap(solver.kernel.compute_elementwise)((x, y), (x, y))
+        expected_state = RPCholeskyState(gramian_diagonal)
+        assert eqx.tree_equal(state, expected_state)
+
+    def test_non_uniqueness(self, reduce_problem: _SupervisedReduceProblem) -> None:
+        """Test that setting `unique` to be false produces no errors."""
+        dataset, _, _ = reduce_problem
+        solver = JointRPCholesky(
+            random_key=self.random_key,
+            coreset_size=10,
+            feature_kernel=self.feature_kernel,
+            response_kernel=self.response_kernel,
+            unique=False,
+        )
+        solver.reduce(dataset)
+
 
 class TestSteinThinning(RefinementSolverTest, ExplicitSizeSolverTest):
     """Test cases for :class:`coreax.solvers.coresubset.SteinThinning`."""
@@ -496,7 +602,7 @@ class TestConditionalKernelHerding(RefinementSolverTest, ExplicitSizeSolverTest)
     """Test cases for :class:`coreax.solvers.coresubset.ConditionalKernelHerding`."""
 
     feature_kernel = SquaredExponentialKernel()
-    response_kernel = SquaredExponentialKernel()
+    response_kernel = PCIMQKernel()
 
     @override
     @pytest.fixture(scope="class")
@@ -516,7 +622,7 @@ class TestConditionalKernelHerding(RefinementSolverTest, ExplicitSizeSolverTest)
         self,
         request: pytest.FixtureRequest,
         solver_factory: Union[type[Solver], jtu.Partial],
-    ) -> _ReduceProblem:
+    ) -> _SupervisedReduceProblem:
         if request.param == "random":
             data_key, supervision_key = jr.split(self.random_key)
             data = jr.uniform(data_key, self.shape)
@@ -525,7 +631,7 @@ class TestConditionalKernelHerding(RefinementSolverTest, ExplicitSizeSolverTest)
             expected_coreset = None
         else:
             raise ValueError("Invalid fixture parametrization")
-        return _ReduceProblem(
+        return _SupervisedReduceProblem(
             SupervisedData(data=data, supervision=supervision), solver, expected_coreset
         )
 
@@ -540,12 +646,14 @@ class TestConditionalKernelHerding(RefinementSolverTest, ExplicitSizeSolverTest)
         scope="class",
     )
     def refine_problem(
-        self, request: pytest.FixtureRequest, reduce_problem: _ReduceProblem
+        self,
+        request: pytest.FixtureRequest,
+        reduce_problem: Union[_ReduceProblem, _SupervisedReduceProblem],
     ) -> _RefineProblem:
         """
         Pytest fixture that returns a problem dataset and the expected coreset.
 
-        An expected coreset of 'None' implies the expected coreset for this solver and
+        An expected coreset of `None` implies the expected coreset for this solver and
         dataset combination is unknown.
 
         We expect the '{well,under,over}-sized' cases to return the same result as a
@@ -585,7 +693,7 @@ class TestConditionalKernelHerding(RefinementSolverTest, ExplicitSizeSolverTest)
 
     # pylint: disable=duplicate-code
     def test_greedy_cmmd_state(self, reduce_problem: _SupervisedReduceProblem) -> None:
-        """Check that the cached ConditionalKernelHerding state is as expected."""
+        """Check that the cached `ConditionalKernelHerdingState` is as expected."""
         dataset, solver, _ = reduce_problem
         solver = cast(ConditionalKernelHerding, solver)
         _, state = solver.reduce(dataset)
@@ -623,7 +731,7 @@ class TestConditionalKernelHerding(RefinementSolverTest, ExplicitSizeSolverTest)
 
     # pylint: enable=duplicate-code
 
-    def test_batching(self, reduce_problem: _ReduceProblem) -> None:
+    def test_batching(self, reduce_problem: _SupervisedReduceProblem) -> None:
         """Test that batching produces no errors."""
         dataset, _, _ = reduce_problem
         solver = ConditionalKernelHerding(
@@ -635,8 +743,8 @@ class TestConditionalKernelHerding(RefinementSolverTest, ExplicitSizeSolverTest)
         )
         solver.reduce(dataset)
 
-    def test_non_uniqueness(self, reduce_problem: _ReduceProblem) -> None:
-        """Test that setting 'unique' to be false produces no errors."""
+    def test_non_uniqueness(self, reduce_problem: _SupervisedReduceProblem) -> None:
+        """Test that setting `unique` to be false produces no errors."""
         dataset, _, _ = reduce_problem
         solver = ConditionalKernelHerding(
             random_key=self.random_key,
@@ -647,7 +755,9 @@ class TestConditionalKernelHerding(RefinementSolverTest, ExplicitSizeSolverTest)
         )
         solver.reduce(dataset)
 
-    def test_approximate_inverse(self, reduce_problem: _ReduceProblem) -> None:
+    def test_approximate_inverse(
+        self, reduce_problem: _SupervisedReduceProblem
+    ) -> None:
         """Test that using an inverse approximator produces no errors."""
         dataset, _, _ = reduce_problem
         solver = ConditionalKernelHerding(