+++ CHANGELOG.md
@@ -10,9 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added Kernelised Stein Discrepancy divergence in `coreax.metrics.KSD`.
 - Added Conditional Maximum Mean Discrepancy (CMMD) metric in `coreax.metrics.CMMD`
+- Added coresubset algorithm to target CMMD `coreax.solvers.coresubset.ConditionalKernelHerding`
 - Added capability to take tensor products of kernels in `coreax.kernel.TensorProductKernel`
 - Added Joint Maximum Mean Discrepancy (JMMD) metric in `coreax.metrics.JMMD`
-- Added coresubset algorithms targeting the joint density in `coreax.solvers.coresubset.JointKernelHerding`
+- Added coresubset algorithms targeting the joint density in `coreax.solvers.coresubset.JointRPCholesky`
+`coreax.solvers.coresubset.JointKernelHerding`
 
 ### Fixed